+++ packages/react-components/react-drawer/stories/Drawer/DrawerPreventClose.stories.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import { Drawer } from '@fluentui/react-drawer';
+import { Drawer, DrawerBody, DrawerHeader, DrawerHeaderTitle } from '@fluentui/react-drawer';
 import { Button } from '@fluentui/react-components';
+import { Dismiss24Regular } from '@fluentui/react-icons';
 
 export const PreventClose = () => {
   const [open, setOpen] = React.useState(false);
@@ -8,10 +9,24 @@ export const PreventClose = () => {
   return (
     <div>
       <Drawer position="right" open={open} lightDismiss={false}>
-        <Button appearance="outline" onClick={() => setOpen(false)}>
-          Close
-        </Button>
-        <p>This drawer cannot be closed when clicking outside nor using the "ESC" key</p>
+        <DrawerHeader>
+          <DrawerHeaderTitle
+            action={
+              <Button
+                appearance="subtle"
+                aria-label="Close"
+                icon={<Dismiss24Regular />}
+                onClick={() => setOpen(false)}
+              />
+            }
+          >
+            Prevent close with Esc or outside click
+          </DrawerHeaderTitle>
+        </DrawerHeader>
+
+        <DrawerBody>
+          <p>This drawer cannot be closed when clicking outside nor using the "ESC" key</p>
+        </DrawerBody>
       </Drawer>
 
       <Button appearance="primary" onClick={() => setOpen(true)}>