+++ packages/react-components/react-tags/src/components/Tag/useTag.tsx
@@ -1,8 +1,9 @@
 import * as React from 'react';
-import { getNativeElementProps, resolveShorthand } from '@fluentui/react-utilities';
+import { getNativeElementProps, resolveShorthand, useEventCallback, useId } from '@fluentui/react-utilities';
 import { DismissRegular, bundleIcon, DismissFilled } from '@fluentui/react-icons';
 import type { TagProps, TagState } from './Tag.types';
-import { useARIAButtonShorthand } from '@fluentui/react-aria';
+import { Delete, Backspace } from '@fluentui/keyboard-keys';
+import { useTagGroupContext_unstable } from '../../contexts/TagGroupContext';
 
 const tagAvatarSizeMap = {
   medium: 28,
@@ -27,13 +28,32 @@ const DismissIcon = bundleIcon(DismissFilled, DismissRegular);
  * @param ref - reference to root HTMLElement of Tag
  */
 export const useTag_unstable = (props: TagProps, ref: React.Ref<HTMLElement>): TagState => {
+  const { dismissible: contextDismissible, handleTagDismiss, size: contextSize } = useTagGroupContext_unstable();
+
+  const id = useId('fui-Tag', props.id);
+
   const {
     appearance = 'filled-lighter',
     disabled = false,
-    dismissible = false,
+    dismissible = contextDismissible,
     shape = 'rounded',
-    size = 'medium',
+    size = contextSize,
+    value = id,
   } = props;
+
+  const handleClick = useEventCallback((ev: React.MouseEvent<HTMLButtonElement>) => {
+    props.onClick?.(ev);
+    if (!ev.defaultPrevented) {
+      handleTagDismiss?.(ev, value);
+    }
+  });
+
+  const handleKeyDown = useEventCallback((ev: React.KeyboardEvent<HTMLButtonElement>) => {
+    props?.onKeyDown?.(ev);
+    if (!ev.defaultPrevented && (ev.key === Delete || ev.key === Backspace)) {
+      handleTagDismiss?.(ev, value);
+    }
+  });
 
   return {
     appearance,
@@ -45,30 +65,34 @@ export const useTag_unstable = (props: TagProps, ref: React.Ref<HTMLElement>): T
     size,
 
     components: {
-      root: 'div',
-      content: 'div',
+      root: 'button',
       media: 'span',
       icon: 'span',
       primaryText: 'span',
       secondaryText: 'span',
-      dismissButton: 'button',
+      dismissIcon: 'span',
     },
 
-    root: getNativeElementProps('div', {
+    root: getNativeElementProps('button', {
       ref,
       ...props,
+      id,
+      onClick: handleClick,
+      onKeyDown: handleKeyDown,
     }),
 
-    content: resolveShorthand(props.content, { required: true }),
     media: resolveShorthand(props.media),
     icon: resolveShorthand(props.icon),
-    primaryText: resolveShorthand(props.primaryText, { required: true }),
-    secondaryText: resolveShorthand(props.secondaryText),
-    dismissButton: useARIAButtonShorthand(props.dismissButton, {
-      required: props.dismissible,
+    primaryText: resolveShorthand(props.primaryText, {
+      required: true,
       defaultProps: {
-        disabled,
-        type: 'button',
+        children: props.children,
+      },
+    }),
+    secondaryText: resolveShorthand(props.secondaryText),
+    dismissIcon: resolveShorthand(props.dismissIcon, {
+      required: dismissible,
+      defaultProps: {
         children: <DismissIcon />,
       },
     }),