+++ packages/react-components/react-toast/src/index.ts
@@ -1,4 +1,4 @@
 export { Toaster } from './components/Toaster';
 
 export { useToastController } from './state';
-export type { ToastPosition, ToastId } from './state';
+export type { ToastPosition, ToastId, ToastOffset } from './state';