+++ packages/react-components/react-toast/src/state/types.ts
@@ -3,7 +3,7 @@ import { EVENTS } from './constants';
 export type ToastId = string;
 export type ToasterId = string;
 
-export type ToastPosition = 'top-right' | 'top-center' | 'top-left' | 'bottom-right' | 'bottom-center' | 'bottom-left';
+export type ToastPosition = 'top-right' | 'top-left' | 'bottom-right' | 'bottom-left';
 
 export interface ToastOptions {
   toastId: ToastId;
@@ -13,12 +13,22 @@ export interface ToastOptions {
   pauseOnWindowBlur: boolean;
   pauseOnHover: boolean;
   toasterId: ToasterId | undefined;
+  priority: number;
+  dispatchedAt: number;
 }
 
+export interface ToastOffsetObject {
+  horizontal?: number;
+  vertical?: number;
+}
+
+export type ToastOffset = Partial<Record<ToastPosition, ToastOffsetObject>> | ToastOffsetObject;
+
 export interface ToasterOptions
-  extends Pick<ToastOptions, 'position' | 'timeout' | 'pauseOnWindowBlur' | 'pauseOnHover'> {
-  offset?: number[];
+  extends Pick<ToastOptions, 'position' | 'timeout' | 'pauseOnWindowBlur' | 'pauseOnHover' | 'priority'> {
+  offset?: ToastOffset;
   toasterId?: ToasterId;
+  limit?: number;
 }
 
 export interface Toast extends ToastOptions {
@@ -31,11 +41,11 @@ export interface CommonToastDetail {
   toasterId?: ToasterId;
 }
 
-export interface ShowToastEventDetail extends Partial<ToastOptions>, CommonToastDetail {
+export interface ShowToastEventDetail extends Partial<Omit<ToastOptions, 'dispatchedAt'>>, CommonToastDetail {
   toastId: ToastId;
 }
 
-export interface UpdateToastEventDetail extends Partial<ToastOptions>, CommonToastDetail {
+export interface UpdateToastEventDetail extends Partial<Omit<ToastOptions, 'dispatchedAt'>>, CommonToastDetail {
   toastId: ToastId;
 }