+++ packages/react-components/react-label/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui/react-label
 
-This log was last generated on Fri, 12 May 2023 20:27:23 GMT and should not be manually modified.
+This log was last generated on Thu, 18 May 2023 00:39:09 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.1.12](https://github.com/microsoft/fluentui/tree/@fluentui/react-label_v9.1.12)
+
+Thu, 18 May 2023 00:39:09 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-label_v9.1.11..@fluentui/react-label_v9.1.12)
+
+### Patches
+
+- Bump @fluentui/react-utilities to v9.9.0 ([PR #27834](https://github.com/microsoft/fluentui/pull/27834) by beachball)
+- Bump @fluentui/react-jsx-runtime to v9.0.0-alpha.4 ([PR #27834](https://github.com/microsoft/fluentui/pull/27834) by beachball)
 
 ## [9.1.11](https://github.com/microsoft/fluentui/tree/@fluentui/react-label_v9.1.11)